Move CI check related functions to their own file
